+++ app.js
@@ -1,8 +1,9 @@
 import express from "express";
 import morgan from "morgan";
 import dotenv from "dotenv";
-import connectDB from "./src/config/database.js";
 import productRoutes from "./src/routes/ProductRoutes.js";
+import articleRoutes from "./src/routes/ArticleRoutes.js";
+import commentRoutes from "./src/routes/CommentRoutes.js";
 import cors from "cors";
 
 const app = express();
@@ -10,22 +11,13 @@ const app = express();
 // 환경변수 설정
 dotenv.config();
 
-connectDB()
-  .then(() => {
-    console.log("mongoDB 연결 성공");
-  })
-  .catch((err) => {
-    console.error("mongoDB 연결 실패", err);
-    process.exit(1);
-  });
-
 // 포트 설정
 app.set("port", process.env.PORT || 8000);
 
 // CORS 설정
 app.use(
   cors({
-    origin: ["http://localhost:5173", "https://panda-market-kr.netlify.app"],
+    origin: [process.env.CORS_ORIGIN_DEV, process.env.CORS_ORIGIN_PROD],
     credentials: true,
   })
 );
@@ -34,6 +26,8 @@ app.use(
 app.use(express.json());
 app.use(morgan("dev"));
 app.use(productRoutes);
+app.use(articleRoutes);
+app.use(commentRoutes);
 
 app.get("/hello", (req, res) => {
   res.status(200).send("Hellod World");