+++ src/controllers/productController.js
@@ -1,22 +1,31 @@
-import Product from "../models/Product.js";
+import prisma from "../config/database.js";
 
 const getProducts = async (req, res, next) => {
   try {
-    const { page = 1, pageSize, orderBy, keyword } = req.query;
+    const { page = 1, pageSize = 10, orderBy = "recent", keyword = "" } = req.query;
     const skip = (page - 1) * pageSize;
-    const sortOption = { createdAt: orderBy === "recent" ? -1 : 1 };
+    const take = Number(pageSize);
+    // Prisma 정렬 옵션 설정
+    const sortOption = orderBy === "recent" ? { createdAt: "desc" } : { createdAt: "asc" };
     const searchQuery = keyword
       ? {
           $or: [{ name: { $regex: keyword, $options: "i" } }, { description: { $regex: keyword, $options: "i" } }],
         }
       : {};
-    const totalCount = await Product.countDocuments(searchQuery);
 
-    const sortedProducts = await Product.find(searchQuery)
-      .sort(sortOption)
-      .limit(pageSize)
-      .skip(skip)
-      .select("_id name price createdAt");
+    const totalCount = await prisma.product.count({ where: searchQuery });
+    const sortedProducts = await prisma.product.findMany({
+      where: searchQuery,
+      skip,
+      take,
+      orderBy: sortOption,
+      select: {
+        id: true,
+        name: true,
+        price: true,
+        createdAt: true,
+      },
+    });
 
     res.status(200).json({
       products: sortedProducts,
@@ -30,7 +39,17 @@ const getProducts = async (req, res, next) => {
 const getProductById = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const product = await Product.findById(id).select("_id name description price tags createdAt");
+    const product = await prisma.product.findUnique({
+      where: { id },
+      select: {
+        id: true,
+        name: true,
+        description: true,
+        price: true,
+        tags: true,
+        createdAt: true,
+      },
+    });
     res.status(200).json(product);
   } catch (error) {
     next(error);
@@ -40,7 +59,7 @@ const getProductById = async (req, res, next) => {
 const createProduct = async (req, res, next) => {
   try {
     const { name, description, price, tags } = req.body;
-    const newProduct = await Product.create({
+    const newProduct = await prisma.product.create({
       name,
       description,
       price,
@@ -56,16 +75,15 @@ const updateProduct = async (req, res, next) => {
   try {
     const { id } = req.params;
     const { name, description, price, tags } = req.body;
-    const updatedProduct = await Product.findByIdAndUpdate(
-      id,
-      {
+    const updatedProduct = await prisma.product.update({
+      where: { id },
+      data: {
         name,
         description,
         price,
         tags,
       },
-      { new: true }
-    );
+    });
     res.status(200).json(updatedProduct);
   } catch (error) {
     next(error);
@@ -75,7 +93,7 @@ const updateProduct = async (req, res, next) => {
 const deleteProduct = async (req, res, next) => {
   try {
     const { id } = req.params;
-    await Product.findByIdAndDelete(id);
+    await prisma.product.delete({ where: { id } });
     res.status(204).send();
   } catch (error) {
     next(error);