+++ src/data/seed.js
@@ -1,24 +1,21 @@
-import mongoose from "mongoose";
 import mockData from "./mock.js";
-import Product from "../models/Product.js";
-import dotenv from "dotenv";
-
-dotenv.config();
+import prisma from "../config/database.js";
 
 const seedDB = async () => {
   try {
-    await mongoose.connect(process.env.DATABASE_URL);
     // 기존 데이터 삭제
-    await Product.deleteMany({});
+    await prisma.product.deleteMany();
 
     // mock 데이터 삽입
-    await Product.insertMany(mockData);
+    await prisma.product.createMany({
+      data: mockData,
+    });
 
     console.log("데이터베이스 시딩 완료");
   } catch (error) {
     console.error("데이터베이스 시딩 실패", error);
   } finally {
-    mongoose.disconnect();
+    await prisma.$disconnect();
   }
 };